Added the Chirp.razor
Added the Chirp.razor by copying the template we were given. Fixed it so we get cheeps, from our database instead of random static cheeps. The cheeping functionality is not ready, since the UI is not ready, but the methods for cheeping to the database are mostly ready, they are not used by anything right now though.

It works for the database on localhost, but the path will have to be changed, when azure is up and running.