Added page identification and suport for interacting with local host
This commit should maybe have been 2 different ones as the changes are a little big.
I have added that the razer wep aplication knowes what page nr should be shown, as this was part of the issue requirements. These page numberes are still not being used yet.

I have also added suport for the "make start-razer-local" command that starts the razer wep app but connected to the localhost servises insted of the one on azure. This is done to help development where there has been made a change to services.cs but you would not like to push that to azure, but razer still needs these changes